refactor(admin): rename credit grants admin area to "Invoices"
Rename the admin section to "Invoices" (what the page actually is):

- Route moved /admin/credit-grants -> /admin/invoices (incl. /new and
  /[invoiceId]); all in-app links, router pushes, and revalidatePath
  targets updated.
- Components moved src/components/admin/credit-grants -> .../admin/invoices
  and imports updated.
- Labels: page title, admin overview section, and breadcrumb segment now
  read "Invoices".
