+++ apps/web/messages/en.json
@@ -265,7 +265,7 @@
       "admin": "Admin",
       "edit": "Edit",
       "enterprise-contracts": "Enterprise contracts",
-      "credit-grants": "Credits",
+      "invoices": "Invoices",
       "mcp": "Model Context Protocol (MCP)",
       "chat": "Chat",
       "tasks": "Tasks",
@@ -2611,7 +2611,7 @@
         }
       },
       "CreditGrants": {
-        "title": "Credits",
+        "title": "Invoices",
         "description": "Grant one-time credits to a user or organization via a Stripe invoice. Marking the invoice paid grants the credits instantly.",
         "cardTitle": "New credit grant",
         "newGrant": "New credit grant",
@@ -2709,7 +2709,7 @@
             "description": "Create and manage draft, active, and canceled enterprise contracts."
           },
           "creditGrants": {
-            "title": "Credits",
+            "title": "Invoices",
             "description": "Grant one-time credits to an organization via a Stripe invoice."
           }
         }
