+++ apps/web/src/app/(app)/admin/invoices/page.tsx
@@ -2,12 +2,12 @@ import type { Metadata } from "next";
 import Link from "next/link";
 import { getTranslations } from "next-intl/server";
 
-import { CreditGrantInvoiceList } from "@/components/admin/credit-grants/credit-grant-invoice-list";
+import { CreditGrantInvoiceList } from "@/components/admin/invoices/credit-grant-invoice-list";
 import { Button } from "@/components/ui/button";
 import { creditGrantAdminService } from "@/lib/services/credit-grant-admin.service";
 
 export const metadata: Metadata = {
-  title: "Credits",
+  title: "Invoices",
   description: "Grant one-time credits to a user or organization",
 };
 
@@ -26,7 +26,7 @@ export default async function CreditGrantsPage() {
             <p className="text-muted-foreground text-sm">{t("description")}</p>
           </div>
           <Button asChild>
-            <Link href="/admin/credit-grants/new">{t("newGrant")}</Link>
+            <Link href="/admin/invoices/new">{t("newGrant")}</Link>
           </Button>
         </div>
